Adding unit test for HTTP headers
pac4cli currently has two "bugs" in its behavior:

1. All HTTP headers are turned to lower-case after passing through
pac4cli.
2. Repeated HTTP headers become missing.

Regarding the first "bug", the HTTP headers are supposed to be
case-insensitive. Still, I've discovered that PHP 5.4 cli server (which
should never be deployed in production) misbehaves if the HTTP headers
are not in the expected case. That issue seems fixed in PHP 7.

So, the first "bug" is not really a bug... But it violates the
Robustness principle: "Be conservative in what you send, be liberal in
what you accept."

Regarding the second "bug", it is really a bug. Duplicate keys for HTTP
headers are valid, and should either be preserved, or combined using
comma-separated values. Either way, the order must be preserved.

Disclaimer: I'm not an expert on HTTP protocol, so I what I'm describing
might deviate from the specification.